Subscription system added for user on purchase

+++ app/Http/Controllers/BkashPaymentController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\BkashTransection;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Inertia\Inertia;
 use Karim007\LaravelBkash\Facade\BkashRefund;
 use Karim007\LaravelBkash\Facade\BkashPayment;
@@ -12,8 +13,13 @@ class BkashPaymentController extends Controller
 {
     private $amount = 0;
     private $payment_id;
+    private $package_id;
     public function index(Request $request)
     {
+        // dd($request->all());
+        $this->package_id = $request->id;
+        session()->put('package_id', intval($request->id));
+        // dd($this->package_id, session()->get('package_id'));
         if (auth()->user()->hasRole('school')) {
             if (session()->has('invoice_amount')) {
                 session()->forget('invoice_amount');
@@ -66,22 +72,45 @@ class BkashPaymentController extends Controller
 
     public function executePayment(Request $request)
     {
-        if (auth()->user()->hasRole('school')) {
+        $user = auth()->user();
+        if ($user->hasRole('school')) {
             $tr = BkashTransection::where('payment_id', $request->paymentID)->first();
-            // dd(session()->get('payment_id'), $request->paymentID);
             $paymentID = $request->paymentID;
             $e = BkashPayment::executePayment($paymentID);
+            // dd($e);
 
-            if ($e['transactionStatus'] !== 'Completed') {
-                $tr->delete();
+            if (isset($e['errorCode']) && $e['errorCode'] !== null) {
+                //show error message on payment failure'
+                dd($e);
             } else {
-                $tr->update([
-                    'payment_id' => $e['paymentID'],
-                    'transaction_status' => $e['transactionStatus'],
-                    'customer_msisdn' => $e['customerMsisdn'],
-                    'trx_id' => $e['trxID'],
-                    'transaction_reference' => $e['transactionReference'] ?? null,
-                ]);
+                if (isset($e['transactionStatus']) && $e['transactionStatus'] !== 'Completed') {
+                    $tr->delete();
+                } else {
+                    DB::transaction(function () use ($user, $tr, $e) {
+                        $s = $user->subscription();
+                        // Update user subscription
+                        if ($s == null) {
+                            $s->create([
+                                'package_id' => session()->get('package_id'),
+                                'will_expire' => now()->addMonth(12),
+                            ]);
+                        } else {
+                            $s->update([
+                                'package_id' => session()->get('package_id'),
+                                'will_expire' => now()->addMonth(12),
+                            ]);
+                        }
+
+                        // Update transection info
+                        $tr->update([
+                            'payment_id' => $e['paymentID'],
+                            'transaction_status' => $e['transactionStatus'],
+                            'customer_msisdn' => $e['customerMsisdn'],
+                            'trx_id' => $e['trxID'],
+                            'transaction_reference' => $e['transactionReference'] ?? null,
+                        ]);
+                    });
+                }
             }
             return $e;
         }
